Initial check-in
This is the implementation of the 'Botz' framework as authored by Aznidin Zainuddin, as found in an archive that is made available on the forums of the Ignite Realtime community.

See https://discourse.igniterealtime.org/t/botz-internal-bot-library-for-openfire/ for the original release notification of this framework.

The sources have been modified slightly. Notably:
- the license has been changed from GPL to Apache2 (to match a similar change for all Ignite projects, years ago)
- the package name has been changed to `org.igniterealtime` (and no longer references Jive Software)
- a basic Maven project framework has been added
- Openfire's deprecated Log implementation has been replaced with slf4j's.

The version number of this import has been based on 1.1.0, in an attempt to not conflict with earlier releases that appear to have been versioned 1.0.0.