Denote release 1.1.0
This is a release that is intentionally not optimized or significantly changed, in an effort to remain as close to the original code for at least one release.

Notable changes between this release and the original code:
- the license has been changed from GPL to Apache2 (to match a similar change for all Ignite projects, years ago)
- the package name has been changed to `org.igniterealtime` (and no longer references Jive Software)
- a basic Maven project framework has been added
- Openfire's deprecated Log implementation has been replaced with slf4j's.

The version number of this import has been based on 1.1.0, in an attempt to not conflict with earlier releases that appear to have been versioned 1.0.0.